Lista 3 - Exerc 5 (final)

+++ Lista3A-Exe5/src/Vereador.java
@@ -56,7 +56,7 @@ public class Vereador {
 		if (this.qtdProjApres == 0) {
 			return 0;
 		}
-		return (this.qtdProjAprov / this.qtdProjApres) * this.getIndiceTrabalho();
+		return (this.qtdProjAprov / (double)this.qtdProjApres) * this.getIndiceTrabalho();
 	}
 
 	public double getIndiceTrabalho() {
